requirements.txt
google-auth
google-auth-oauthlib
google-cloud-storage
